Fix a bug in external_command.rb
Under some circumstances, closing a file descriptor can
cause an error. The case I saw was a TCP socket that would
raise an "Uninitialized stream" error if you called .close
on it. This would exit the child process, leaving it as a
zombie with the parent blocked in select().

Therefore ignore errors from close().