Improve the _mysociety_commonlib_directory function
We had a bug report from an Alaveteli user, which on investigation
turned out to be caused by the fact that cd was producing output
in the _mysociety_commonlib_directory function. The only possible
reason for that I can see is that CDPATH might be set in the environment,
which could also under some circumstances cause the function as
was to return the wrong answer, so we unset CDPATH and also direct
the output from cd to /dev/null just in case there is some other
even-more-obscure combination of circumstances that could cause
cd to produce output even with CDPATH unset.