Add a generic script for installing mySociety sites
There was a lot of shared code between the install scripts
for FixMyStreet and MapIt.  This script contains that common
code, and can be extended to install any mySociety site.

This script should be run as root, for instance with:

  sudo ./install-site.sh --default fixmystreet fms

... which will clone the fixmystreet repository, and source
the site-specific-install.sh script in its bin directory
to complete the installations.

As with the previous versions of these scripts, this should
only be used on a clean server install, since it overwrites
several configurations files and makes significant changes
to your server's setup.