install-site.sh: preserve the script and arguments as invoked
On EC2 instances that are set up to use the default domain, we may
need to write a new hostname into various configuration files on
reboot - this is important when booting up an instance based on an
AMI, as well.  On FixMyStreet, for example, this is relatively simple,
and there is an ec2-rewrite-conf script in the repository that can be
used.  With other sites this is more involved, and it's simpler to
just re-run the install script with the same arguments.  In order to
do the latter, we need to preserve the script as used to install the
site in the first place, and the arguments it was invoked with.  This
commit introduces a script to do that recording, and alters
overwrite_rc_local to reinstall if ec2-rewrite-conf isn't present.