install-site.sh: attempt to set Postfix's mydestinations
As far as I can tell, this is the correct way to set the
mydestinations line in Postfix's main.cf on installation
of the Debian package.  However, this doesn't work, for
reasons that aren't clear to me, and we have to fix it up
later, so I've left this in with a FIXME comment in the
hope that we can investigate it further when there is
more time.