Prefer ./script over ./bin
It is common for Ruby applications to have both a ./bin and ./script
directory.

This commit forces install-site.sh to prefer ./script over ./bin if both
exist, as this is where site-specific-install.sh lives in alaveteli.

We now only attempt to source the file if $BIN_DIRECTORY exists, and
fail gracefuly otherwise.