Pin mySociety packages for Wheezy only
debian.mysociety.org contains a) packages missing from Squeeze and b)
newer versions of packages that are not available in the official
Squeeze repos.

The latter case is the issue here.

On Wheezy, we want to lower the priority because we don't want to
install an older version from debian.mysociety.org in preference to the
official Wheezy package.

On Squeeze, we _do_ want to take the version from debian.mysociety.org
because it will generally contain newer or bug-fixed versions.