install-site.sh: On Ubuntu trusty don't create template_postgis
On Ubuntu 14.04 (Trusty Tahr) the script for creating template_postgis
doesn't work, and PostGIS support should be added via CREATE EXTENSION
instead. Any project that uses install_postgis and has been updated to
use this version of commonlib should be updated; for example the
corresponding changes for MapIt are here:

  https://github.com/mysociety/mapit/pull/156

Also fix the version of the Django script used to one that still
supports Debian squeeze.