Use open4 to handle forking and closing file handles
Our homegrown code had problems when applying a timeout to an external
process - the database handle of the parent process could disappear and
cause any transaction in process to be rolled back. In this
implementation open4 handles the low level business of closing other
file handles on fork, and doesn't seem to have the same issue.