[install-site] Initial support for Docker builds
Adds a `--docker` parameter. This will set a handful of other variables
(`INSTALL_DB`, `INSTALL_POSTFIX` and `DEFAULT_SERVER`) that are used to
control installation of separate components.

Brief notes added to the help text and a sanity check on the `SITE`
argument as for now this is only supported by FixMyStreet.