Add support for a `--systemd` option to install scripts (#61)
This adds support for using a native systemd unit file template when
installing services if the `--systemd` option is provided and there
is a file present at `$CONF_DIRECTORY/systemd.example`.

It also amends the regex to replace the `USER` line in the sysvinit
templates to work with `User=` lines in systemd unit files as well.