Updates to supported distributions in install scripts
This removes support for everything before Xenial and Stretch and adds
support up to Focal and Buster.

It should be noted that the mySociety repository only has formal support
for Debian and not Ubuntu, so each Ubuntu release is currently mapped to
the closest available Debian version, which may work for at least some
packages.

This also refactors the repository sections slightly, templating them
based on the specific versions in question to reduce duplication.