Add GNU Privacy Guard to core packages
Required for installing on Bullseye. Without this, for Alaveteli, we see
an error:
```
  Updating mySociety APT source...
  E: gnupg, gnupg2 and gnupg1 do not seem to be installed
```

We can't add to our site specific packages as these installed after the
repos have been updated.

While we could install prior to the install script invocation in our
Vagrantfile this wouldn't allow production installs and it seems like
other sites will also require this to get pass APT update.