Switch default database connection sslmode from allow to prefer
An `sslmode` of `allow` will first try an unencrypted connection before
falling back to an encrypted one should that fail. `prefer` is the
opposite behaviour and a more secure default, so let's use that instead.

Ideally this would be configurable, but with this change any connection
errors logged by PostgreSQL will indicate that SSL isn't configured
which is more useful information that indicating that it is enforced.