[MapIt] Don’t hardcode ";" as query param separator
Using the semicolon as a separator is disabled in recent
versions of Python such as those used on the new
infrastructure.

This change is quite conservative as I’m not confident
it doesn’t have knock-on effects elsewhere. Sites that
need to can opt-in to using the standard ‘&’ separator
by setting it as MAPIT_API_SEPARATOR the config value.