Update APT configuration to conform with newer deb882 format
This updates our APT source definitions to use the newer deb882 format
and renames `/etc/apt/sources.list` to `/etc/apt/sources.list.dist` if it
exists to avoid duplicate warnings when running commands.

In addition, this also removes the use of `apt-key` in favour of storing
the signing key for our repository in `/etc/apt/keyrings` (which will be
created if this doesn't exist) and pointing at that via `Signed-By` in
the sources, removing deprecation warnings from `apt-key`.

See: https://github.com/mysociety/commonlib/issues/87