[PHP] Standardise breaking from `switch` in IsterSimpleXMLElement.php
Using `continue` generates a warning we're not skipping the rest of the
`foreach` loop but only jumping out of the `switch` block as we would if
we used `break`.

To explicitly move to the next iteration of the containing loop, we
should use `continue 2`. However in this case it is academic as there is
no more code after the `switch`, so I'm using `break` for consistency
with the rest of the code.